+++ app/services/domain/CommonService.scala
@@ -320,9 +320,9 @@ object CommonService extends AbstractService {
     BigDecimal(value).setScale(2, BigDecimal.RoundingMode.HALF_UP).toDouble
   }
 
-  def getOthers(orgs: Array[(String, Int)], total: Int): Int ={
+  def getOthers(orgs: Array[(String, Int)], total: Long): Int ={
 
-    total - orgs.map(x=>x._2).sum
+    total.toInt - orgs.map(x=>x._2).sum
     //percentInfor(orgQueries,total)
   }
